#18: rename file/folder names and component names
