+++ waste-lookup/src/components/Cockpit/SearchBar/SearchBar.js
