+++ waste-lookup/src/components/SearchResults/SearchResults.css
